apps/ml-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
numpy==2.1.3
scikit-learn==1.5.2
xgboost==2.1.3
joblib==1.4.2
psycopg[binary]==3.2.3
pgvector==0.3.6
sentence-transformers==3.3.1
datasets==3.1.0
boto3==1.35.71
mlflow==2.18.0
